requirements.txt
agno
fastapi
uvicorn
transformers
torch
nest-asyncio
anthropic
openai
psycopg[binary,pool]
prompt_toolkit
haystack-ai
qdrant-haystack
neo4j
neo4j-haystack
sentence-transformers
tabulate
einops
loguru
gradio
qdrant-client
markdown-it-py
mdit_plain
nltk
sqlalchemy
pytest
duckduckgo-search